+++ ui/desktop/src/goosed.ts
@@ -1,14 +1,33 @@
 import path from 'node:path';
-import { spawn } from 'child_process';
+import { execSync, spawn } from 'child_process';
 import { existsSync } from 'fs';
 import log from './utils/logger';
 import os from 'node:os';
+import { createServer } from 'net';
+import { loadZshEnv } from './utils/loadEnv';
 
-// Start the goosed binary
-export const start = (app) => {
-  // In development, the binary is in src/bin
-  // In production, it will be in the resources/bin directory
+
+// Find an available port to start goosed on
+export const findAvailablePort = (): Promise<number> => {
+  return new Promise((resolve, reject) => {
+    const server = createServer();
+
+
+    server.listen(0, '127.0.0.1', () => {
+      const { port } = server.address() as { port: number };
+      server.close(() => {
+        log.info(`Found available port: ${port}`);
+        resolve(port);
+      });
+    });
+  });
+};
+
+// Goose process manager. Take in the app, port, and directory to start goosed in.
+export const startGoosed = async (app, dir=null): Promise<number> => {
+  // In will use this later to determine if we should start process
   const isDev = process.env.NODE_ENV === 'development';
+
   let goosedPath: string;
 
   if (isDev) {
@@ -18,59 +37,60 @@ export const start = (app) => {
     // In production, use the path relative to the app resources
     goosedPath = path.join(process.resourcesPath, 'bin', process.platform === 'win32' ? 'goosed.exe' : 'goosed');
   }
+  const port = await findAvailablePort();
 
-  log.info(`Starting goosed from: ${goosedPath}`);
+  // in case we want it
+  //const isPackaged = app.isPackaged;
   
-  // Check if the binary exists
-  if (!existsSync(goosedPath)) {
-    log.error(`goosed binary not found at path: ${goosedPath}`);
-    return;
-  }
-
+  // we default to running goosed in home dir - if not specified
   const homeDir = os.homedir();
-
-  // Optional: Ensure the home directory exists (it should, but for safety)
-  if (!existsSync(homeDir)) {
-    log.error(`Home directory does not exist: ${homeDir}`);
-    return;
+  if (!dir) {
+    dir = homeDir;
   }
 
-  console.log("Starting goosed in: ", homeDir);
-
+  log.info(`Starting goosed from: ${goosedPath} on port ${port} in dir ${dir}` );
+  
   // Define additional environment variables
   const additionalEnv = {
     // Set HOME for UNIX-like systems
     HOME: homeDir,
     // Set USERPROFILE for Windows
     USERPROFILE: homeDir,
+
+    // start with the port specified 
+    GOOSE_SERVER__PORT: port,
   };
 
   // Merge parent environment with additional environment variables
   const env = { ...process.env, ...additionalEnv };
 
   // Spawn the goosed process with the user's home directory as cwd
-  const goosedProcess = spawn(goosedPath, [], { cwd: homeDir, env: env });
-
+  const goosedProcess = spawn(goosedPath, [], { cwd: dir, env: env });
 
   goosedProcess.stdout.on('data', (data) => {
-    log.info(`goosed stdout: ${data.toString()}`);
+    log.info(`goosed stdout for port ${port} and dir ${dir}: ${data.toString()}`);
   });
 
   goosedProcess.stderr.on('data', (data) => {
-    log.error(`goosed stderr: ${data.toString()}`);
+    log.error(`goosed stderr for port ${port} and dir ${dir}: ${data.toString()}`);
   });
 
   goosedProcess.on('close', (code) => {
-    log.info(`goosed process exited with code ${code}`);
+    log.info(`goosed process exited with code ${code} for port ${port} and dir ${dir}`);
   });
 
   goosedProcess.on('error', (err) => {
-    log.error('Failed to start goosed:', err);
+    log.error(`Failed to start goosed on port ${port} and dir ${dir}`, err);
   });
 
   // Ensure goosed is terminated when the app quits
+  // TODO will need to do it at tab level next
   app.on('will-quit', () => {
     log.info('App quitting, terminating goosed server');
     goosedProcess.kill();
   });
+
+  return port;
 };
+
+
