+++ ui/desktop/src/main.ts
@@ -2,7 +2,7 @@ import 'dotenv/config';
 import { loadZshEnv } from './utils/loadEnv';
 import { app, BrowserWindow, Tray, Menu, globalShortcut, ipcMain, Notification } from 'electron';
 import path from 'node:path';
-import { start as startGoosed } from './goosed';
+import { findAvailablePort, startGoosed } from './goosed';
 import started from "electron-squirrel-startup";
 import log from './utils/logger';
 import { getPort } from './utils/portUtils';
@@ -30,9 +30,9 @@ const checkApiCredentials = () => {
 };
 
 let appConfig = { 
-  GOOSE_SERVER__PORT: 3000, 
+  apiCredsMissing: !checkApiCredentials(),
   GOOSE_API_HOST: 'http://127.0.0.1',
-  apiCredsMissing: !checkApiCredentials()
+  GOOSE_SERVER__PORT: 0,
 };
 
 const createLauncher = () => {
@@ -81,7 +81,9 @@ const createLauncher = () => {
 let windowCounter = 0;
 const windowMap = new Map<number, BrowserWindow>();
 
-const createChat = (query?: string) => {
+const createChat = async (app, query?: string) => {
+
+  const port = await startGoosed(app);  
   const mainWindow = new BrowserWindow({
     titleBarStyle: 'hidden',
     trafficLightPosition: { x: 16, y: 18 },
@@ -94,7 +96,7 @@ const createChat = (query?: string) => {
     icon: path.join(__dirname, '../images/icon'),
     webPreferences: {
       preload: path.join(__dirname, 'preload.js'),
-      additionalArguments: [JSON.stringify(appConfig)],
+      additionalArguments: [JSON.stringify({ ...appConfig, GOOSE_SERVER__PORT: port })],
     },
   });
 
@@ -195,36 +197,21 @@ const showWindow = () => {
 
 app.whenReady().then(async () => {
   // Load zsh environment variables in production mode only
-  const isProduction = app.isPackaged;
-  loadZshEnv(isProduction);
-
-  // Get the server startup configuration
-  const shouldStartServer = (process.env.VITE_START_EMBEDDED_SERVER || 'yes').toLowerCase() === 'yes';
   
-  if (shouldStartServer) {
-    log.info('Starting embedded goosed server');
-    const port = await getPort();
-    process.env.GOOSE_SERVER__PORT = port.toString();
-    appConfig = { ...appConfig, GOOSE_SERVER__PORT: process.env.GOOSE_SERVER__PORT };
-    startGoosed(app);
-  } else {
-    log.info('Skipping embedded server startup (disabled by configuration)');
-  }
-
   createTray();
-  createChat();
+  createChat(app);
 
   // Show launcher input on key combo
   globalShortcut.register('Control+Alt+Command+G', createLauncher);
 
   app.on('activate', () => {
     if (BrowserWindow.getAllWindows().length === 0) {
-      createChat();
+      createChat(app);
     }
   });
 
   ipcMain.on('create-chat-window', (_, query) => {
-    createChat(query);
+    createChat(app, query);
   });
 
   ipcMain.on('notify', (event, data) => {
